Use requestAnimationFrame to throttle DOM patches.
- Use rAF to throttle renders / batch render calls.
- Introduce new `init` event that run before the view
  has rendered the first time, but after the app internal
  state has been initialized.
  - Revert `ready` to `loaded` event. This event now fires
    once when the view has been attached to the DOM.

OVERVIEW

Think of requestAnimationFrame (rAF) as a clever setTimeout.

rAF tells the browser to call a specified function before the
next screen repaint. This allow us to throttle DOM patches in
our render pipeline.

Imagine your app calls hundreds (or thousands) of actions within
a very short time span, a likely scenario in real world apps.

This is how the story unfolds. We:

  (1) Call your action to obtain a new/partial state and merge
  it with the global state (or return immediately if the action
  doesn't return anything). At the end of this step we have a new
  global state.

  (2) Call render() (which calls your view() function to get the
  vtree). The vtree is likely to be large and complex, composed
  of hundreds or thousands of h() calls.

  (3) Call patch() with the new and old vtree. Patch walks through
  the new vtree comparing each vnode with its corresponding vnode
  in the old vtree and patches the real DOM where it needs. Patch
  essentially means creating, updating, replacing or removing DOM
  nodes. This is usually the bottleneck.

  (4) Finish. Step out of app.

How can we improve (2) and (3) other than by writing a more
efficient patch algorithm?

> TODO: Write a more efficient patch algorithm.

INRODUCING rAF

What's going to happen now is that after (1) we are going to
lock the render() call.

    function render() {
      if (!lockRender) {
        lockRender = true

        requestAnimationFrame(() => {
          PATCH
          lockRender = false
        })
      }
    }

This mechanism allow us to immediately skip over to (4) and not
block,i.e., not call view() and patch().

Actions will continue to run skipping (2) and (3) straight-to (4).

The global state will continue to be updated and action/update
events will continue to fire.

At some point, and in good timing, the browser will run the
function we scheduled using rAF. It's then, and only then when
(2) and (3) will be finally called.

Finally, inside the callback, we now can unlock the render()
function to allow the scheduling of future patches.